requirements.txt
# requirements
pyinstaller
Pillow
pymsgbox
six